refactor: my own way of structuring directories
