+++ src/app/tasks/tasks.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { TaskService } from './task.service';
+import { TasksService } from './tasks.service';
 
-describe('TaskService', () => {
-  let service: TaskService;
+describe('TasksService', () => {
+  let service: TasksService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(TaskService);
+    service = TestBed.inject(TasksService);
   });
 
   it('should be created', () => {
